refactor(SelectMenu): Eliminate ActionMenu Dependencies
