+++ src/components/ActionMenu/ActionMenu/ActionMenu.tsx
@@ -4,7 +4,7 @@ import { AriaList } from "@components/ActionMenu/Aria/AriaList";
 import { AriaMenuItem } from "@components/ActionMenu/Aria/AriaMenuItem";
 import { AriaSection } from "@components/ActionMenu/Aria/AriaSection";
 import { getDisabledItemIds, getKeyItemRecord, getMenuItems, mapToAriaProps } from "@components/ActionMenu/Aria/helper";
-import { MenuItemType } from "@components/SelectMenu/SelectMenu";
+import { MenuItemType } from "@components/Dropdown/SelectMenu/SelectMenu";
 import { useMenu } from "@react-aria/menu";
 import { useTreeState } from "@react-stately/tree";
 import { FocusStrategy } from "@react-types/shared";
