+++ src/components/ActionMenu/index.ts
@@ -1,1 +1,1 @@
-export * from "../SelectMenu/SelectMenu";
+export * from "../Dropdown/SelectMenu/SelectMenu";
