+++ src/components/Dropdown/SelectMenu/SelectMenuItem.tsx
@@ -6,7 +6,7 @@ import { Node } from "@react-types/shared";
 import { FOCUS_STYLE_INSET } from "@utilities/focusStyle";
 import { merge } from "@utilities/merge";
 import React, { FC, useRef } from "react";
-import { MenuItem } from "../MenuItem";
+import { MenuItem } from "../../MenuItem";
 import { MenuItemType } from "./SelectMenu";
 
 type SelectMenuItemProps = {
